Update configure to check for pre-existing files
Check that pre-existing configs exist before attempting
to move them to prevent an error message.

+++ configure.sh
@@ -6,6 +6,9 @@ LINKS=".bbtools .xsession .xinitrc .pekwm .i3lock .wbar"
 
 for LINK in $LINKS
 do
-    mv ~/$LINK ~/${LINK}.${DATE}
+    if [ -f ~/$LINK ]
+    then
+        mv ~/$LINK ~/${LINK}.${DATE}
+    fi
     ln -s ~/git/config/$LINK ~/$LINK
 done
